Migrated vwap indicator to the class pattern.
