+++ js/masters/indicators/vwap.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/VWAPIndicator.js';
+import '../../Stock/Indicators/VWAP/VWAPIndicator.js';
