+++ ts/Stock/Indicators/VWAP/VWAPPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type VWAPIndicator from './VWAPIndicator';
+import type SMAPoint from '../SMA/SMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class VWAPPoint extends SMAPoint {
+    public series: VWAPIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default VWAPPoint;
